Bootstrap empty solution and project
This commit creates a brand new solution that contains a single project.

The project will contain the implementation of the
`Microsoft.UI.UIAutomation` API. It's configured as a simple empty DLL
project that targets a pre-release version of the Windows SDK, given
that this is needed to get access to the platform
`Windows.UI.UIAutomation` APIs that the project will provide utilities
for.

The solution will contain all projects that are associated to the main
utility library, including tests and demo apps.