Add Microsoft.UI.UIAutomation.idl and configure build steps
This commit adds the `Microsoft.UI.UIAutomation.idl` file that describes
the current `Microsoft.UI.Automation` API. It configures the MIDL
compiler to build this as WinRT metadata and adds a custom build step
that invokes the `cppwinrt.exe` tool to generate component
implementation code for the WinRT components described by this metadata.

Beyond producing the `winmd` metadata file and the auto-generated C++
source files, this commit doesn't add any functionality to the DLL
itself.