Add the UiaOperationAbstraction static lib
The `UiaOperationAbstraction` is a C++ static library that can help C++
consumers switch between a mode where calls are added and executed as
part of a Remote Operation or simply executed one after the other
(classic UIA). Switching between the two modes is as simple as setting a
bool.

This library is not only specific only for C++ consumers, but even for
C++ clients it might not always be the best choice. It is useful for
consumers that expect to need to switch between "classic" UIA and remote
operations-based calls. It allows such clients to express the same
algorithm using the same exact code and then switch between those modes
of execution at will. However, there is no free lunch, and this brings
in both runtime overhead, as well as cognitive overhead for dealing with
some of the abstractions that this library introduces.