Add support for remote operation cache requests
Cache requests in remote operations behave slightly differently than in
classic UIA. Rather than being tied to a specific operation (e.g.
GetParentElementBuildCache), a remote element's cache can be populated
at any time. This is because we are already in the remote process, so we
don't need to attach a cache request to an outgoing operation to reduce
overall latency.

The instructions IsCacheRequest, NewCacheRequest, PopulateCache,
CacheRequestAddProperty, and CacheRequestAddPattern are added. The WinRT
API provides a fairly thin wrapper on top of them.

The UiaOperationAbstraction library is a bit more complex. Since it
should work with both remote and classic UIA, we need to follow a
similar interface. This is done by introducing the UiaCacheRequest type.
This type cannot be imported or exported from a remote operation, but it
can be constructed from either an IUIAutomationCacheRequest or an
AutomationRemoteCacheRequest. In methods that accept cache requests, we
then check if we're in a remote operation. If we are, we call
PopulateCacheRequest using the cache request's remote form. If we're not
remote, we retrieve the underlying COM pointer and pass it on to UIA as
usual.