Enable Azure Pipelines build validation (#11)
In order to enable automatic CI in Azure Pipelines, we need to ensure that the appropriate version of the Insider Preview SDK is installed on the build machine. This is achieved by the new `Install-WindowsSdkISO` script. This is the same approach that other projects use, for instance `microsoft-ui-xaml`. The script itself is reused from MUX. 

The PR also fixes an issue with build order/dependency for the C# demo app: it now explicitly depends on the `Microsoft.UI.UIAutomation` project (since it needs the metadata produced by this step). The `Windows.winmd` file that was referenced when `Microsoft.UI.UIAutomation.winmd` itself was built is copied to the same output directory just to make sure that all projects wind up referencing all the same metadata.


Co-authored-by: Marko Lalic <markola@microsoft.com>