Add UiaChar and UiaString indexing (#13)
This commit adds an implementation of `UiaChar` -- a wrapper around
`Microsoft.UI.UIAutomation.AutomationRemoteChar` in the same style as
all other "abstraction" wrappers. The local equivalent of this type is
`wchar_t`. This required the general plumbing for converting results to
local values as well as converting local values to remote operands.

This new `UiaChar` type is then used to implement `UiaString::At`, a
method that allows character-level indexing of `UiaString`. The ability
to index remote strings is supported by the UIA platform (Remote
Operations VM) as well as the builder library, so this is just an
addition to the abstraction library.

Co-authored-by: Marko Lalic <markola@microsoft.com>