UiaHwnd: Add implementation of conversion operator to UIA_HWND (#18)
The UiaHwnd class has the conversion operator declared in the header file, which aims to be used to get to the local type from the abstraction type. But the implementation is not in place, so the client of operation abstraction such as Narrator will get the linking error while building DLLs since the implementation is missing in the static lib.

This change adds the simple implementation here to get the local type.