Add support for `Stringify` operations to UIA operation abstraction (#34)
**Remote Operations** provide a variety of types that can be used to perform more complex calculations on the provider side. The types range from "simple" numeric types, collections (arrays, maps) to UIA proprietary types (elements, text ranges).

UIA clients often times have to work with the UIA tree, analyze it and perform operations based on relationships between elements in the tree. However, there is no "simple" way of creating and remembering the relationships/associations in **Remote Operations** today.

That is because Remote Operations provide 2 different collection types:
1. Flat arrays,
2. Maps from strings to any object type (supported by UIA).

(2) would be ideal to remember relationships between elements. However, UIA elements identify themselves through `RuntimeId`s that are arrays of `int`s and therefore cannot be used as keys of maps in **Remote Operations**.

This change adds supports for the `Stringify` instruction and adds tools for the instruction to the UIA operation abstraction library to let the consumers of the library use `RuntimeId`s (and other properties, types) as keys of `UiaStringMap`s to create and use (non-flat, tree-like) object associations in **Remote Operations**.

The change comes with tests to cover local and remote usage of the new operation (via the UIA operation abstraction library).

Fixes #33

Co-authored-by: Lukasz Makar <lukamak@microsoft.com>