WinRT Builder: Add GetMetadataValue support (#36)
This PR adds support for `GetMetadataValue` to the WinRT Builder library.

The first part of the PR is adding the `AutomationMetadata` enum to the WinRT surface area. This is done by running the (currently still closed-source) code generator to generate all the relevant pieces for an enum based on the information in the classic COM typelib.

The `GetMetadataValue` method added to `AutomationRemoteElement` is then a simple wrapper around the underlying platform Remote Operations VM bytecode instruction, which follows the existing pattern of builder wrappers.

A functional end-to-end test is added, but only for a provider that doesn't support metadata currently (Xaml). The (currently still closed-source) larger test suite in the OS does validate the underlying instruction against providers that do support it. Such a test should be added to this repo once we've migrated the full test suite into the open source repo.

Partially addresses #35 -- C++ Abstraction work is left to be added in a separate PR.