Add an ability to multi-bind input and result variables via UIA operation abstraction (#31)
Currently, when using the UIA operation abstraction library, inputs and results can only be bound to the operation scope one by one. While the functionality is absolutely sufficient to define remote operations with any number of input arguments and output/result values, it creates unnecessarily long, repeating patterns of chains of `BindInput` and `BindResult` calls.

The change adapts the proposed (in the issue) solution for it by adding variadic `BindInput` and `BindResult`. It also modifies one of the tests to use the new version of `BindResult`.

Fixes #5 

Co-authored-by: Lukasz Makar <lukamak@microsoft.com>