Fix `UiaEnum` comparison operators for all types `UiaEnum` can be constructed out of (#39)
`UiaEnum` is a wrapper type that can be constructed out of the following types:
1. COM `enum`s such as `CONTROLTYPEID` (local type),
2. WinRT `enum`s of the operation builder such as `AutomationControlType` (local type),
3. Corresponding stand-in (numeric) type (remote type),
4. `UiaEnum` itself (copy-construction).

The abstraction class is providing 3 different types of comparison methods against the mentioned values depending on what the caller has. In short, `UiaEnum` allows comparing against COM, WinRT `enum`s as well as the `UiaEnum` type (the abstraction tries to hide the corresponding stand-in type).

However, this does not work today due to ambiguity of defined comparison operators (and which should be applied for which situation/type).

In fact, the only comparison operators that work are the ones for numeric values (such as `CONTROLTYPEID` cast to `long` in case of `UiaControlType`) and WinRT `enum`s (such as the `AutomationControlType` `enum class`). But these comparators cannot be applied when the code author tries to write UIA operation abstraction-based functions that take UIA abstraction types as the functions usually cannot assume whether abstraction types hold "local" or "remote" types (and therefore casting `UiaEnum`s is not an option).

This change attempts to fix that problem by:
1. Keeping the comparison operators for comparing `UiaEnum`s,
2. Reduces other operators to template-based functions that take a value to compare a `UiaEnum` against, create a `UiaEnum` out of the value and then compares the two (using the kept operators).
3. Removing any other defined comparison operators for the COM and WinRT `enum` types as well as the `rvalue` comparison types.

With the change, ensuring type correctness is pushed towards `UiaEnum` constructors since the operators create `UiaEnum` instances out of `rhs` before comparing values.

Fixes #38

Co-authored-by: Lukasz Makar <lukamak@microsoft.com>