Add support for `UiaTuple` to the UIA operation abstraction library (#43)
## Background

Currently, UIA operation abstraction and its wrappers give access to:
1. Primitive types (`int`, `bool`, `uint`, `double`, character, ...),
2. Variants,
3. Enums (`COM`, `WinRT`),
4. UIA types (element, patterns, rectangle...),
5. Strings,
6. Some collections (arrays, maps from string to other types).

The library does not provide support for any `struct`-like types. The closest it gets to that is through collections of variants.

This change attempts to add support for `UiaTuple` to mimic C++'s `std::tuple` and to provide more tools to work with types that provide access to multiple fields of different value types.

## Changes
First and foremost, the change adds `UiaTuple` to the collection of types that the library offers. As with other wrappers, the class is a variant of two types:
1. Local, which is `std::shared_ptr<std::tuple<LocalTypes...>>`
2. Remote, which is `AutomationRemoteArray` (as it is the case for `UiaArray`).

The class offers functionalities specific to `std::tuple` such as:
1. Strongly typed `Set` by the statically provided index of the field to access and set,
2. Strongly typed `Get` by the statically provided index of the field to access and return,
3. Comparison operators (`==` and `!=`),
4. Construction out of local types (similar to `UiaInt value{ 0 }`),
5. Construction out of wrapper types for convenience to enable constructs like `UiaTuple<UiaElement, UiaRange> elementRange{ uiaElement, uiaRange }`.

**Note:** Right now, `UiaTuple` does not implement `Stringify`.

The change adds a set of tests to cover basic usage and functionalities for `UiaTuple`, which include:
1. Constructing instance of the class in different ways (default, with local values or by passing abstraction wrappers),
2. Getting and setting values for the tuples,
3. Using `UiaTuple` objects as values of collections.

In addition to that, the change adds support for "empty" `UiaVariant` values to enable default initialization of variant wrappers.

**Note:** Some tests intentionally use wrapper methods instead of local type methods (for instance `UiaStringMap` methods over `std::map` methods) even after resolving **Remote Operation** results to force the compiler to generate and analyze abstraction wrapper methods and check their correctness (since the code is template-based and therefore require usage to be instantiated and compiled).

Fixes #44

Co-authored-by: Lukasz Makar <lukamak@microsoft.com>