Add any cast in element and pattern class (#54)
The UiaStringMap is a template class that support different wrapper types as the value of map. Since this is a template class, the corresponding instantiated class used as value need to implement all necessary function or variable to pass the compile time. The LookUp function in UiaStringMap requires the class used as value has function pointer c_anyCast variable defined so that we could cast the AutomationRemoteAnyObject to specific value class object. Unfortunately, classes in our generated files, such as UiaElement, are missing c_anyCast defined which resulted in the compiling errors when UiaStringMap used to hold class like UiaElement as value.

This PR fixes the issue by adding missed defined variable.