Update enum and metadata wrapper (#59)
The reason for making this change is to make the wrapper supports get metadata value functionality.
Therefore, a helper is added `UiaVariant GetMetadataValue(UiaPropertyId propertyId, UiaMetadata metadataId);`
Once we get the returned `UiaVariant`, the proper usage is to cast this to the exact enum type we want (for example `UiaSayAsInterpretAs`). However, this cast is not supported in the wrapper yet. Also, `UiaVariant::IsNull()` is added, otherwise cannot really use the returned value.

To address the cast issue from `UiaVariant` to `UiaEnum`, an update is made to the `UiaEnum` definition to include a `CastFunc` template parameter. The type of this function is defined as:
`StandinT (winrt::Microsoft::UI::UIAutomation::AutomationRemoteAnyObject::*)() const`.
After adding `CastFunc`, we could assign the cast function for each `UiaEnum` class. And assign to `c_anyCast`, which will let us utilize `UiaVariant::AsType` to cast.

For example, `UiaSayAsInterpretAs` could specify the cast function as `&winrt::Microsoft::UI::UIAutomation::AutomationRemoteAnyObject::AsSayAsInterpretAs` to convert a `UiaVariant` to `UiaSayAsInterpretAs`

The type definition in `UiaTypeAbstractionEnums.g.h` is updated based on template parameter change.

Note that, in the `UiaVariant::AsType` function, we have to add an extra static cast for the local value. This is not required before because the returned local type always matches the type of the local variant. But now we need this static cast to convert local variant with `VT_I4` type to a local enum type (for example `winrt::Microsoft::UI::UIAutomation::AutomationSayAsInterpretAs`).