UiaOperationAbstraction:  allow converting a UiaElement to a UiaVariant and back again (#71)
UiaOperationAbstraction has a UiaVariant type for storing data of many types. This is very useful inconjunction with the UiaArray or UiaStringMap types, allowing for holding collections of data of varying types.
Although all the basic data types (UiaInt, UiaString, UiaBool etc) are supported by UiaVariant, UiaElement is not. Meaning that it is impossible to store a UiaElement in a UiaArray or UiaStringMap of variants.
 
this pr allows constructing a UiaVariant given a UiaElement as its constructor argument. UiaVariant also now exposes an IsElement method for checking if the variant is holding a UiaElement, and an AsElement method, which allows for fetching the UiaElement as a UiaElement.
 
This work involved:
* Adding the appropriate helper members to UiaElement for partispating in UiaVariant conversion such as: 
 * c_comVariantType, set to VT_UNKNOWN (IUnknown)
 * c_variantMember, pointing to VARIANT::punkVal
 * c_anyTest, pointing to AutomationRemoteAnyObject::IsElement
 * c_anyCast was already there.
* Adding an explicit constructor for UiaVariant which takes a UiaElement
* adding IsElement and HasElement methods to UiaVariant.

I have also added a function test that successfully tests constructing a UiaVariant from a UIAElement, varifying that IsElement is True, fetching the element with HasElement, and verifying it is a functioning and correct UiaElement by checking its name property.
 
In the process of writing this pr, I did need to forward declare UiaElement in UiaOperationAstraction.h, as the UiaElement name is required now by UiaVariant. I'm not sure if this is okay, or there is a cleaner way to do this?

Also I'd appreciate careful review of the UiaVariant constructor that takes a UiaElement, especially the bit handling construction of the local type:
```
UiaVariant(details::MakeVariantFrom<UiaElement>(wil::com_ptr<IUIAutomationElement>(value.get()).detach())))
```
The existing coding style here shoving everything into the member initializer makes for hard reading.