Add IsOpcodeSupported support to WinRT and C++ wrappers (#78)
IsOpcodeSupported is an API provided by CoreAutomationRemoteOperation
which determines whether a given remote operation opcode is supported by
the current provider connection. Which opcodes are supported depends on
both the client and provider UIA versions.

This change simply plumbs IsOpcodeSupported through
Microsoft.UI.UIAutomation and UiaOperationAbstraction. A more elegant
design taking advantage of the characteristics of each wrapper can be
made in the future.

For UiaOperationAbstraction, we always return true from
IsOpcodeSupported in the local case.

Note, IsOpcodeSupported will throw E_FAIL on future versions of Windows
when no connection has yet been established. Once such versions of
Windows are released, we should update tests and comments with the
specifics. This is tracked by issue #77.