Allow tests to successfully run on Windows 11 (handling new Calculator UI) (#90)
## problem
The majority of the functional tests for the microsoft-ui-uiAutomation project use the Calculator modern app, in-built to Windows.  But due to changes to the Calculator's UI on Windows 11, several of the tests fail, making it impossible to successfully run all tests on Windows 11.
Specifically, the Calculator app now has an extra ancestor element (with a custom controlType) between the grouping element and the window element.
 
## Solution
Changed the affected tests to detect and skip over the new Custom element if found.

## Testing performed
Successfully ran all functional tests on Windows 11 build 22572, Calculator version 11.2201.4.0.